Update back-end part Add DELETE end-point Add routes for the Books endpoint Add CORS settings
